+++ admin/includes/view_all_posts.php
@@ -24,6 +24,35 @@
                     confirmQuery($delete_post_query);
                     break;
                 
+                case 'clone':
+                    $query = "SELECT * FROM posts WHERE post_id = {$postValueId}";
+                    $select_post_query = mysqli_query($connection, $query);
+
+                    while($row = mysqli_fetch_array($select_post_query)){
+                        $post_title = $row['post_title'];
+                        $post_category_id = $row['post_category_id'];
+                        $post_author = $row['post_author'];
+                        $post_status = $row['post_status'];
+                        $post_image = $row['post_image'];
+                        $post_tags = $row['post_tags'];
+                        $post_content = $row['post_content'];
+
+                    }
+
+                    $query = "INSERT INTO posts(post_category_id, post_title, post_author, 
+                        post_date, post_image, post_content, post_tags,  post_status) ";
+
+                        $query .= "VALUES ('$post_category_id', '$post_title', '$post_author', now(),
+                        '$post_image', '$post_content', '$post_tags', '$post_status')";
+
+                        $copy_query = mysqli_query($connection, $query);
+
+                        if(!$copy_query){
+                            die("QUERY FAILED!".mysqli_error($connection));
+                        }
+
+                    break;
+                
                 default:
             }
         }
@@ -41,6 +70,7 @@
                 <option value="published">Publish</option>
                 <option value="draft">Draft</option>
                 <option value="delete">Delete</option>
+                <option value="clone">Clone</option>
             </select>
         </div>
         <div class="col-xs-4">
@@ -59,6 +89,7 @@
                 <th>Image</th>
                 <th>Tags</th>
                 <th>Comments</th>
+                <th>Views</th>
                 <th>Date</th>
                 <th>Edit</th>
                 <th>Delete</th>
@@ -66,7 +97,7 @@
         </thead>
         <tbody>
             <?php
-        $query = "SELECT * FROM posts";
+        $query = "SELECT * FROM posts ORDER BY post_id DESC";
         $select_post = mysqli_query($connection, $query);
         while ($row = mysqli_fetch_assoc($select_post)) {
             $post_id = $row['post_id'];
@@ -78,6 +109,7 @@
             $post_tags = $row['post_tags'];
             $post_comment_count = $row['post_comment_count'];
             $post_date = $row['post_date'];
+            $post_views_count = $row['post_views_count'];
             
             echo "<tr>";
             ?>
@@ -103,8 +135,9 @@
             echo "<td>{$post_tags}</td>";
             echo "<td>{$post_comment_count}</td>";
             echo "<td>{$post_date}</td>";
+            echo "<td><a href= 'posts.php?reset={$post_id}'>{$post_views_count}</a></td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this post?');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
             echo "</tr>";
         }
         
@@ -120,6 +153,14 @@ if(isset($_GET['delete'])){
     $query = "DELETE FROM posts WHERE post_id = {$post_id_del} ";
     $delete_query = mysqli_query($connection, $query);
     confirmQuery($delete_query);
+    header("Location: posts.php");
+}
+
+if(isset($_GET['reset'])){
+    $post_id_del = $_GET['reset'];
+    $query = "UPDATE posts SET post_views_count = 0 WHERE post_id = ". mysqli_real_escape_string($connection,$_GET['reset']) ." ";
+    $reset_query = mysqli_query($connection, $query);
+    confirmQuery($reset_query);
     header("Location: posts.php");
 }
